old-tests/submission/Microsoft/semantic/mark.html

<!DOCTYPE html>
<html><head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type">
  <title>verify MARK is a HTMLElement</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 <link href="http://www.w3.org/TR/html5/the-mark-element.html#the-mark-element" rel="help">
</head>
<body>
<div id="log"></div>
<p>This is just an example of what MARK can do:</p>
<pre> <code>
 var i: Integer;
 begin
  i := <mark id="mark">1.1</mark>;
 end.
 </code>
</pre>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("mark").toString() )
 })
</script>


</body></html>
